MANIFEST.in ensures docs and extension get built
I've added a MANIFEST.in file that recursively adds docs/ and
mozmill/extension/ to the build.  Without it, on my system, neither are
included in the build or installation, likely because of the recent
change in revision control systems.

When the Mozmill revision control system was Subversion, Python's
setuptools was able to automatically infer which files needed to be
added to the mozmill.egg-info/SOURCES.txt file, and hence the build.  It
did this by including directories that contained .svn folders, reasoning
that if the folder was under revision control, then it was likely
important, and should be included.

With the switch to the Git, the .svn folders no longer exist.  As of
this writing, and with my current setup, setuptools is no longer able to
infer which files should be included in the build, so I added the
MANIFEST.in file.